// Frontend/pets-ui/src/app/models/pet.model.ts
export interface Pet {
  id: number;
  name: string;
  breed: string;
  age: number;
  sex: string;
  size: string;
  color: string;
  weight: number;
  imageUrl: string;
  healthStatus: string;
  vaccinationStatus: string;
  dewormed: boolean;
  neuteredSpayed: boolean;
  medicalNotes: string;
  temperament: string;
  houseTrained: boolean;
  adoptionFee: number;
  location: string;
  availabilityStatus: string;
  intakeDate: string;
  description: string;
  trainingLevel: string;
  contactInfo: string;
}

export interface PetFormData {
  name: string;
  breed: string;
  age: number;
  sex: string;
  size: string;
  color: string;
  weight: number;
  imageUrl: string;
  healthStatus: string;
  vaccinationStatus: string;
  dewormed: boolean;
  neuteredSpayed: boolean;
  medicalNotes: string;
  temperament: string;
  houseTrained: boolean;
  adoptionFee: number;
  location: string;
  availabilityStatus: string;
  intakeDate: string;
  description: string;
  trainingLevel: string;
  contactInfo: string;
}